span-processor: Move span map to a separate file
